CAS-609: CAP Voting Options Page (#651)
* update style

* split into more components

* updates

* add class

* create component

* extract into components

* keep alignment for create community

* fix header spacing

* add class

* add proposal title on letNavPannel

* update

* update style

* update space style

* update style

* updates

* add Header

* updates

* updates

* fix component name

* fix component name

* comment imports

* fix component name

* update

* add paddinh

* updates

* remove comment

* rename component

* rename component

* updates

* add class

* remove padding and border for mobile

* make header scrollable

* hide back button

* update

* update

* remove from import

* split form and add second step

* extract into component

* updates

* update

* remove span and fix check

* remove borders on mobile

* add fiend

* fix validation

* fix padding bottom

* remove output

* this file should be on nex PR
