+++ frontend/packages/client/src/App.sass
@@ -229,6 +229,32 @@ label > input[type="radio"].is-chosen + *::before, .is-button-chosen
 	background-repeat: no-repeat
 label > input[type="radio"].is-chosen + *::before
 	background-color: white
+label > input[type="radio"].green-radio
+	display: none
+	-webkit-appearance: none
+	appearance: none
+	background-color: #fff
+	margin: 0
+	font: inherit
+	color: $grey-lighter
+	width: 20px
+	height: 20px
+	border: 1px solid $grey-lighter
+	border-radius: 50%
+	display: grid
+	place-content: center
+label > input[type="radio"].green-radio::before
+	content: ""
+	width: 14px
+	height: 14px
+	border-radius: 50%
+	transform: scale(0)
+	transition: 120ms transform ease-in-out
+	box-shadow: inset 1em 1em $success
+label > input[type="radio"].green-radio:checked::before
+	transform: scale(1)
+label > input[type="radio"].green-radio:checked
+	border-color: $success	
 .proposal-copy
 	color: $grey
 .proposal-copy ul
