+++ frontend/packages/client/src/components/ProposalCreate/StepThree/index.js
@@ -1,0 +1,1 @@
+export { default as StepThree } from './StepThree';
