+++ frontend/packages/client/src/components/ProposalCreate/index.js
@@ -1,3 +1,4 @@
 export { default as PropCreateStepOne } from './StepOne/index';
+export { StepThree as PropCreateStepThree } from './StepThree';
+export { default as PropCreateStepFour } from './StepFour';
 export { StepTwo as PropCreateStepTwo } from './StepTwo';
-export { default as PropCreateStepThree } from './StepThree';
