+++ frontend/packages/client/src/pages/ProposalCreate.js
@@ -6,6 +6,7 @@ import { useModalContext } from 'contexts/NotificationModal';
 import { useWebContext } from 'contexts/Web3';
 import { ErrorModal, StepByStep, WalletConnect } from 'components';
 import {
+  PropCreateStepFour,
   PropCreateStepOne,
   PropCreateStepThree,
   PropCreateStepTwo,
@@ -71,11 +72,12 @@ export default function ProposalCreatePage() {
       return;
     }
 
-    const { strategy, minBalance, maxWeight, name, body } = stepsData[0];
+    const { name, body } = stepsData[0];
+    const { strategy, minBalance, maxWeight } = stepsData[1];
 
     const hasValidStartTime = isStartTimeValid(
-      stepsData[1].startTime,
-      stepsData[1].startDate
+      stepsData[2].startTime,
+      stepsData[2].startDate
     );
 
     if (!hasValidStartTime) {
@@ -87,16 +89,16 @@ export default function ProposalCreatePage() {
     }
 
     const startTime = parseDateToServer(
-      stepsData[1].startDate,
-      stepsData[1].startTime
+      stepsData[2].startDate,
+      stepsData[2].startTime
     ).toISOString();
 
     const endTime = parseDateToServer(
-      stepsData[1].endDate,
-      stepsData[1].endTime
+      stepsData[2].endDate,
+      stepsData[2].endTime
     ).toISOString();
 
-    const choices = stepsData[0].choices.map((c) => ({
+    const choices = stepsData[1].choices.map((c) => ({
       choiceText: c.value,
       choiceImgUrl: c?.choiceImgUrl ?? null,
     }));
@@ -141,17 +143,24 @@ export default function ProposalCreatePage() {
         useHookForms: true,
       },
       {
-        label: 'Set Date & Time',
+        label: 'Voting Options',
         description:
           'Some description of what you can write here that is useful.',
         component: <PropCreateStepTwo />,
         useHookForms: true,
       },
       {
-        label: 'Preview Proposal',
+        label: 'Date and Time',
         description:
           'Some description of what you can write here that is useful.',
         component: <PropCreateStepThree />,
+        useHookForms: true,
+      },
+      {
+        label: 'Preview Proposal',
+        description:
+          'Some description of what you can write here that is useful.',
+        component: <PropCreateStepFour />,
       },
     ],
   };
